Added scaffold line argument to not put connection string in dbcontext
Current database-first workflows require running the `dotnet ef scaffold` command any time database updates have been done so the `DbContext` (and all associated POCOs) reflect the current schema.  

However, the current scaffold command creates a `DbContext` that contains an `OnConfiguring()` method that contains the connection string that was used for context generation.  When running scaffold command after every update it is easy to forget to strip the connection string out prior to committing to source control.  It can also lead to accidental scenarios where the `DbContext` is using this connection string instead of the proper one for its environment (e.g. development database when it should be connecting to production).

To remedy this I added a command line option for the scaffold command of `--dont-add-connection-string`.  When this is provided it scaffolds the `OnConfiguring()` method (so users can see that the override exists) but it does not add the connection string to the generated code.

This will allow users to execute scaffolding without having to worry about sensitive information being checked into source control.
Fixed compile errors

References issue #10432