Add a test helper for creating default service instances
For some types that have a lot of unit tests there is a lot of code just to call the default constructors for a service, its dependency objects, and its dependencies. This change adds a simple helper that uses the D.I. system to create factories for this code. This greatly reduces the amount of test code that needs to change when a new dependency is added. This change uses the factory for type mappers. Use for other types can be added as needed.