Allow RelationalTypeMapper to run in a legacy mode for 2.0 providers
The internals of RelationalTypeMapper have changed significantly to support type conversions. While this was done in a way that should not break 2.0 providers, it turns out that some 2.0 providers take an interesting approach involving throwing exceptions on methods that were not previously called but now can be. To workaround this, if the provider has not been updated--detected by still calling the obsolete constructor--then the RelationalTypeMapper falls back to working in a similar way to how it did before, which should be much less likely to cause issues. This means type conversions is a feature that needs a new provider, but in general that was already the case.

See also #10654