Enable property injection of services, mostly for lazy-loading
Part of issues #10509 #3797

This change introduces a new kind of property: IServiceProperty. These properties are bound to some service of the context, such as ILazyLoader. They can be constructor injected as before, but can also be set as a property both in materialization and on Attach/Detach. When attaching, the service property is set to the service from the current context. When detaching, the service is set to null.

Lazy-loading proxies add an ILazyLoader property to the proxied entity. These means that lazy-loading is disabled when the entity is detached and can then be safely used after the context has been disposed. (It is also possible to disable the context disposed warning for this.) Also, if a proxy is attached to a new context, then it will start doing lazy-loading using that new context.

The service property can also be used when doing lazy-loading without proxies to get the same behavior on detach/attach. In this case it needs to be added to the entity explicitly, but then can be discovered by convention.