Fix to #10785 - Model Query Filter generates incorrect SQL
Problem was that we would not apply null semantics properly to ColumnReference expressions - we were not detecting that the column behind it is actually nullable and IS NULL term needs to be generated for it.
Also as part of this change improved test infrastructure for cases when some of the results are null.