Fix to #10828 - Query: include doesn't work for queries with order by and group by
Problem was that when we try to stream group by, we add order by clause (by grouping key) to the query - this needs to be the first ordering, so that streamed results appear in correct order and can be grouped into buckets. We add this order by during result operator translation (Handle group on QueryModelVisitor).
When the query includes collection, the order by needs to be added to the cloned QM that will fetch the included entities. This however only needs to happen for relational provider, when the grouping is going to be streamed.
Include can't detect that so we either were not putting orderby when needed or putting it when it was not needed, which leads to bad results.

Fix is to add the order by before include, so that when include clones then QM it already has the appropriate ordering, and the ordering will only be added in relational.
Adding order by on the QM level can be problematic for some queries that contain Skip/Take/Distinct operations and customer-specified ordering. For those cases we switch to buffering the groupby, and completely avoid adding the ordering.