Remove ad-hoc enum conversions
Issues #10748, #3620

These are now handled by type converters in type mapping, and for providers that can natively handle enums they will not be converted at all.

Also, change the way properties with conversions are stored in the snapshot such that:
- Only "store types" are used
- A dummy value converter is created to preserve any mapping hints, etc.

Note that if the database maps the enum directly, then this means that it is a "store type" and will be used in the snapshot and migrations and these may break if the type is later removed or changed.