Support change tracking of non-primitive types
Issue #9753

This is primarily focused at allowing providers to generate type mappings for non-primitive type mappings. It can also be used for non-primitive types that are used through value conversions, but #10961 covers adding fluent API for this. Also, this implementation does not attempt to address perf issues, but I added at note to #9422.

I'm not happy with the name ValueComparer--we should discuss in API review.