Fix to #10101 - Query: Select.Include.OrderBy ignores the include
Problem was that in QueryOptimizer, if we flattened redundant subquery (Include result operator was the only element on that subquery and we strip those out) we did not update PathFromQuerySource on the IncludeResultOperator annotations with the new Qsre.
Also, in nav rewrite we were trying to match PathFromQuerySource with regular comparison, rather than using ExpressionComparer.