Fix to #10988 - Query: accessing property on a required navigation defined on a derived type produces INNER JOIN
Problem was that when determining whether given navigation should be represented by INNER JOIN we were looking if it was required and its direction (dependent to principal). However if navigation is defined on derived type in TPH, it can be required and dependent to principal, and should still be represented as LEFT JOIN, because base types don't have the relationship defined for them.

Fix is to check whether navigation is not defined on derived type on top of previous checks.