Query: Parametrize non-EF Queryables for query caching
Issue: Non-EF IQueryables are opaque to us so we cannot compare them causing new cache entry for each of them. But we have mechanism for running a query with parameter of type enumerable using 2nd level cache.
In this case, we can just parametrize non-EF IQueryable to cause a parameter which gets inlined for Contains in Sql

Resolves #11015