Query: Don't funcletize DbSet/InternalDbSet
Issue: A linq query starting from DbSet always has EntityQueryable. But for a manually created query incorrectly passes Constant(IQueryable) rather than IQueryable.Expression then we try to parametrize it and cause multiple running queries.
The fix is to not funcletize it. The Constant(DbSet) gets converted to EntityQueryable during ReLinq query parsing.
This is a negative case. The query is malformed. We would leak context instance in this case. Since it is hot path, we are not going to add overhead for all other cases.

Resolves #11043