Fix to #10440 - Query: Include on derived throws in async
Problem was that for async include on derived entity we produce the following fixup task and await it:

fixup: (QueryContext queryContext | Gear entity | object[] included | CancellationToken ct) =>
{
  return (entity is Officer) ? Task queryContext.QueryBuffer.IncludeCollectionAsync(...) : default(Task)
}

this results in null-value task in case the entity is not of the derived type, and we try to await on it. Fix is to produced Task.CompletedTask instead of default(Task) - we still produce default for non-async cases.

Also added a lot of include tests for async - ported AssertIncludeQuery test helper and most of the gears of war include tests.