Simplify TypeMappingInfo by removing explicit value conversion info
Because providers do not need this info, and based in decisions made for the model snapshot/seed data, should not behave differently based on it.

Instead, the part of the code that knows about converters--in TypeMappingSource--now handles the converters separately, creating appropriate clones of TypeMappingInfo as needed with information from the converter.