Don't mark rowversion as modified if the token hasn't changed
Fixes #9856

The fix is to add a ValueComparer for this mapping so that it does deep comparison. Testing revealed a couple of bugs in ValueCompaer--they were not being used when explicitly setting CurrentValue or OriginalValue.

Providers: This is only done for SQL Server, since this is a SQL Server-specific mechanism, but it is possible other providers have similar issues.