Increase timeout for async database creation tests
Issue:
The test mimics exception for first 5 attempts and passes on 6th attempt. We retry at interval of 1ms. Total duration before timeout is 1 minute.
In case of async test, if the thread is not scheduled to run at least 6 time in 1 min timeout period, we would not be able to reach 6th attempt and hence throw exception.

Resolves #11443