Fix to #11459 - Query: incorrect data returned for queries projecting a single constant from a subquery
Also fixed related issues:
#10946 - ArgumentOutOfRangeException when ordering by a constant in collection include
#11525 - Query: order by anonymous object is silently ignored

Problem was that when a constant was part of the projection, we were never incorporating it into the SQL, but rather materialize it on the client. This is incorrect in case of subquery, since other queries could depend on that constant projection.
Example is include collection with order by a constant - as part of include rewrite we generate a subquery whose projection consists of all the orderings from the outer query, and the collection query is ordered by the elements in the projection. Since before constants were not translated into sql, the resulting query was incorrect, trying to order by element that didnt exist.

Fix is to incorporate projected constants into SQL in case of a subquery. Constant projections of outermost queries are still materialized on the client to avoid transferring large amounts of data over from the database.

Also modified behavior for queries ordering by anonymous object. Before, those queries in some cases those were ignored, in some cases they were producing incorrect queries. Now, we will fall back to client evaluation.