Migrations: Allow moving tables and sequences into default schema
Before this, if the schema/name hadn't changed, we'd generate rename operations with the new schema/name values null. After this, we always generate the operations with the new schema and name. The new behavior distinguishes between moving a table or sequence into the default schema, and not changing the schema. Before, it was ambiguous.

Fixes #8726