Query: TaskLiftingExpressionVisitor should never give null CancellationTokenParameter
Issue: We had an ET generated which used `queryContext.CancellationToken` as CT parameter. We use that property to pass CT from user code to our execution pipeline.
We cannot use that in LambdaExpression since it is not ParameterExpression.
Fix is to use a dummy CT for Lambda since actual CT will be injected from queryContext automatically.


Resolves #11561