Fix to #11622 - Query: Correlated subquery with complex ordering fails during compilation
Problem was that when we were building QM for correlated subquery we relied on query model cloning when trying to remap order by elements, that need to be applied to the correlated collection query.
Relinq cloning is currently broken (https://www.re-motion.org/jira/projects/RMLNQ/issues/RMLNQ-111) if the cloned query contains two copies of the same query source. This is the case if the order by or predicate of the main query contains a subquery itself.

Fix is to not rely on AdjustExpressionAfterCloning to remap complex orderings - now we get them directly from the cloned QM. Only simple orderings, that are guaranteed to be key accesses are remapped.
Also fixed issue in RelationalProjectionExpressionVisitor, where we would visit a projection (which could be complex), add appropriate SQL fragment to SqlExpression, but we would not modify the returned expression, so that the same subquery would be visited later in the pipeline, which would cause unnecessary N+1 queries, or some unnecessary columns being fetched from the database.