Query: Clone inner SubQuery before flattening it out
Issue: When using let keyword in linq, ReLinq would create 1 SubQueryExpression being used in multiple places.
When we are flattening it out, it would mutate the SubQueryExpression so the other references are corrupted now.
This throws Must be reducible node exception since during translation the qsre in the tree is not in the scope of that subquery model

Resolves #11425