Query: Owned entity with non-nullable PK generates invalid SQL when subquery
Issue: For owned Entity when PK is non-nullble (which is sharing column with owner PK), we conclude both the columns to be same. This stops creating proper aliases for subquery generating invalid SQL.
Fix: We de-dupe same columns inside subquery when pushing down. We cannot do the same on top level due to valuebuffer index bindings