Metadata: Bring in symmetry for reference counting for FKs defined on QueryType
Issue: When adding an FK on query type we don't add it to `ReferenceingFKs` collections on principal end.
But when we modify any facet of FK which requires rebuilding FK, we still try to remove FK from those collections which are not initialized hence throws NullRef.
Fix: Only do fixup on such collection when FK is not on QueryType

Also remove HasIndex API from QueryTypeBuilder

Resolves #11720