Fix to #11687 - Query with projection cast on correlated collections throws on 2.1.0-preview2
Problem was that when we apply correlated collection optimization we were not accounting for scenario where the result collection element type was different than the input. This can be achieved by using ToList<T> or ToArray<T> etc.

Fix is to flow the collection result type as well as the input type to the CorrelateSubquery method, and construct the resulting collection based on the output type, if they are different.