Query: Improve alias-ing of projection in SelectExpression
Issue: When we try to find name of the projection, we go through Convert/NullableExpression nodes. But when updating the alias we just wrap it inside a new AliasExpression. If the inner expression had AliasExpression giving it a name, then we would have multiple "AS" in SQL
Fix: When assigning new alias to projection, we run visitor and tries to replace old name to new name to avoid multiple aliases.

Resolves #11757