Query: Fix for GroupByAggregate with join on grouping key produces incorrect sql
Issue: We save `Key` memberInfo to translation mapping in SelectExpression so we can bind `Key` later.
In this query, we also had anonymous projection in aggregate which projected out `Key` which would be essentially same projection but different `MemberInfo`
This causes us to have multiple `MemberInfo`s with same projection.
During `PushDownSubquery()`, we always lifted first `MemberInfo` matching current projection. This caused us to have stale entry in mapping.
In this particular scenario it got used and generated wrong query. (In other cases, we were corrupting mapping but never used it).

Fix:
Correct fix is to remove the stale mapping since the older memberInfo cannot be referenced further. So during the set method, we remove if there is existing mapping to same projection.
This has minor issue in case when GroupBy is done on anonymous object since it causes the inner select to be pushed down inside the elementSelector which requires the older mapping to map correctly. Fix is to update new mapping from key translation only after we have visited elementSelector too.

Resolves #11814