Query: Translate GroupBy new {} to GroupBy constant 1
Resolves #11905

Issue: GroupBy(g => new {}) converts to GroupBy(Constant) because of key not being correlated with range variable.
We translate GroupBy constant to server but in this case, the constant is of anon type which cannot be put in SQL tree hence translation pipeline failed.
Fix is to inject `1` as constant whenever we see GroupBy constant of a type which cannot be represented in SQL.
We also don't save it in mapping so that if key is projected out, we construct it again on client side.