Fix to #11953 - Query with coalescing operator causes exception in 2.1.0-rc1
Problem was that when translating a subquery projecting a non-nullable column that was converted to nullable, the information about nullability is lost - the resulting SelectExpression's type is that of the projected column. This in turn causes problem for the coalesce scenario, which requires the first argument to be nullable.
Fix is to account for possible nullability loss during translation of select expressions which project a single scalar.