Query: Client GroupBy with anonymous type Key causes binding error
Issue: We have custom logic to map the anonymous key in selector to translate key for SQL Group By case which was being applied for client group by causing error during client binding.
Fix for this makes sure that when we do such transformation we are dealing with ShapedQuery only.

Resolves #11973