Fix to #11937 - Null checking in Automapper Project To throws error
Problem was that when addressing issue 11459 we started translating constant values inside subqueries to SQL. This also included constant nulls of non-mapped types, like DTOs - those could get generated as part of null protection/checking or explicitly via cast.

Fix is to only translate constants nulls that we have mapping for (unless they are typed as object). Constant nulls that are not mapped will be evaluated client-side.