Fix to #11923 - Failure to create concrete instance of entity navigation collection class
Problem was that when we construct ClrCollectionAccessor, we would not generate a create delegate if the setter delegate was not available. However for the case where we project a collection navigation without a setter, we need to be able to create the instance of that navigation.

Fix is to produce create delegate whether set delegate was constructed or not.