Always do tracking query for Load
Issue: If the ChangeTracking behaviour is set to NoTracking then calling Load on tracked entity won't track resuls of load query. Which results in load failing.
The fix is to issue AsTracking query so that results from load query are added to state manager and will be fixed up with tracked query

Currently calling Load on non tracked query does not work so this fix is safe in that regard.

Resolves #12209