Add support for defining included column for indexes in SQL server
This commit adds `ForSqlServerInclude` extension to `IndexBuilder` allowing defining columns to be included in a index on sql server.

Generic version of `IndexBuilder` together with its extensions is introduced to support expression parameter.

Breaking change is introduced in `SqlServer` adding a property to interface - `ISqlServerIndexAnnotations.IncludeProperties`.

`MigrationsModelDiffer` now uses structural comparison for annotations to support arrays as annotation value.

Fixes #4846