Fix to #12298 - Query: function call with argument that has null protection applied on can be unnecessarily complex leading to client eval
Problem was that when deciding whether new NullConditional expression was needed we were not correctly comparing the access expressions. Fix is to use ExpressionComparer instead to correctly identify equivalent expressions.
Moreover, when it comes to translation of SqlFragments (e.g. string compare) we are passing the expressions without stripping them from NCEs, and string compare was not accounting for presence of NCEs (or convert nodes). Fix is to improve the translator to correctly match expressions wrapped in NCE.

Also added translator for string.Compare(string, string) which we missed before.