[2.1.3] Query: Do not rewrite entity equality to key equality for complex expression
Issue: In 2.1 we started rewriting entity equality to key equality for complex cases like FirstOrDefault. But when we do Navigation.PK to FK conversion, we actually proper member access path rather than complex expression. Further even if we optimize it to normal key equality, it would still fail if don't translate it to server.
Solution: Do not convert entity equality when its not member access path defined in EF.

Resolves #12256