Fix to #12358 - The use of the ToListAsync method leads to missing data in version 2.1.0 of Microsoft.AspNetCore.All
Problem was in the implementation of TaskLiftingExpressionVisitor and how it interacts with correlated collection optimization. When we perform task lift we do so recursively if we haven't encountered a CancellationToken parameter yet (essentially fishing for it, so that we know what to use).
This becomes a problem if correlated collection optimization async is inside another async method (like Task.FromResult). We ended up recursively lifting the task twice, which resulted in the same correlate call executed twice, which is a problem because both calls share reference to the same reader.

Fix is to fetch CancelationToken parameter as a separate step and don't recursively lift the tasks. It can still deadlock potentially, but will now return correct results.