Add client evaluation not supported exception
It is not always feasible for providers to implement client-side
evaluation for database functions mapped to extension methods on
`DbFunctions`. In these cases, it is common to have the method
throw a `NotSupportedException`.

The `ClientEvaluationNotSupportedException` helps to clarify that
the exception is caused by a function being evaluated rather than
translated.