Allow enum parameters to be passed as integers
Fixes #12557

This is a simple case that used to work for enums because they were always converted to their underlying type anyway. However, it falls out of the space where normal value converters would be expected to work, so to bring back the old behavior we check if the parameter type is already the underlying type for enum conversions and, if so, don't try to convert it.