Fix to #12582 - Includes With Cast Throws Exception
Problem was that when we rewrite collection query during the include compilation we are supposed to prune the result operators that change the result type (e.g. GroupBy). However we missed Cast and OfType, which would lead to discrepancy between the selector and the QM result type.
Fix is to remove the missing operators from the collection query - we only need it to select all the elements of the included collection, we don't need to preserve the original type.