Fix to #12709 - Query: ArgumentOutOfRangeException when doing multiple ordering on same property in collection include
Problem was that for correlated collection, the inner query model projects all orderings for this and all upstream orderbys. However, when we generate SelectExpression, duplicate projections are eliminated. Therefore if we naively construct correlated collection projection list, without taking duplicates into account, this can lead to inconsistencies between SelectExpression and query model.

Fix is to watch for duplicates when constructing the projections for correlated subquery QM and maintain the mapping between orderings and their corresponding projection element, so we know how to remap the orderings correctly - before we just assumed no duplicates and could simply remap them in sequence.