Fix to #13000 - Avoid creating new query roots in navigation rewrite for owned types
Fix is to add a property to NavigationRewritingExpressionVisitor. When the property is set to true owned navigations will not get converted into joins/groupjoins but rather left as they are. Because of this, we can no longer expect source of the navigation rewrite to be QSRE, but rather an arbitrary expression.