WIP: Add eviction to query cache
Issue #12905

I started this with the intention of leaving IMemoryCache in place and adding eviction around it. However, I could not see that IMemoryCache was providing any value other than overhead, so I removed it here.

We have #12157 open to consider changing the way we use IMemoryCache in 3.0, which means that even if we stop using it now, we may want to start using it again in the future.

Therefore, open questions:
* Is it okay to stop using IMemoryCache?
* Should we obsolete the UseMemoryCache methods? After this change they do nothing, but we may want them back in 3.0.
  * Instead of obsoleting, we could update the API doc comments to reflect the current situation
* At the moment, the query limit and number queries to evict is hard-coded. I think it should be possible to configure these in case this change kills perf for some app using a huge number of queries. But then if the implementation changes in 3.0, any configuration we add here may not apply. So ideas:
  * Make them protected properties in the internal service, so people can set them if needed, but only as a workaround to hitting issues.
  * Add proper config and obsolete if needed in 3.0.