[2.1.5] Query: Fix entity equality on navigation accessed on subquery
Issue:
We have certain optimizations running when comparing navigations with null or other navigations
- collection nav == collection nav converts to their parent equality
- collection nav == null converts to parent == null
- dependent to principal nav == null converts to parent.FK == null

For all above cases, we have to create parent expression. We used to create it through a method which assumed the navigation chain will always start from QSRE.
This all worked correctly upto 2.0 where we did not convert equalities after subquery.
In 2.1 we enabled equality rewrites involving subquery, which ultimately broke above assumption. The navigation chain can start from a subquery too.
Which means that for any of above optimization, we would through NRE (or equivalent).

Fix:
Instead of using function which would assume the root of chain is QSRE always, we moved to unwrapping last navigation, exactly same way we discover it in first place.

Resolves #12738