Fix to #12170 - NullReferenceException when using QueryFilter filtering on owned types
Problem was that when we flatten subquery during QM optimization, we clone the querymodel but we were not updating entity mappings for weak types. This resulted in those types not being found later in the pipeline causing the error. Fix is to update the mapping after we clone the query model.