Fix to #11944 - Query: Include for derived types fails when navigation property is defined in an intermediate type
Problem was that when we were constructing collection subquery for the included entities we were using a MemberInfo to determine on which type the navigation is declared. This was incorrect when navigation from EF perspective is defined in a different place than the CLR classes would indicate. We should be using information contained on IPropertyBase instead.