Deep copy byte arrays into the in-memory store
Fixes #13254

This change introduces a new ValueComparer type that can be used when deep copies are needed. Neither the regular or the key comparer can be relied on to work this way, hence the need for a new one. There is no reliable way to know if such a comparer is needed, so it's up to the application/provider setting up the type mapping to provide one if needed.