Remove quirk tests due to global race condition
Fixes #13262

Since two test suites both switch the quirk on and off, but the quirk is global, it results in a race condition where one test is expecting it to be off, but it then gets switched off while that test is running. Fix is to remove these tests since they have little value going forward.