Scaffold NTS literals for in model snapshot
Part of #8741

Currently, as a first step, the scaffolding using WKT as a serialization format. This could be replaced with "pure" C# code calling nested constructors, etc. Also, SRID is not yet round-tripped.

Also, changed to a generic ValueConverter to take advantage of infrastructure and reduce custom overrides in the converter.

Still to do for #8741 - literals in other places