Change Spec.Tests dependency on xunit.core to xunit.extensibility.execution (#13443)
Per xunit guidance, class libraries which deliver xunit bits shouldn't also depend on the xunit or xunit.core metapackage. These packages contain project settings to make a test project runnable, but are not needed for class libraries which need to compile for xunit API's.

> https://xunit.github.io/releases/2.3:
> As a reminder: If you're extending xUnit.net and want to publish your extension as a NuGet package, you should import xunit.extensibility.core and/or xunit.extensibility.execution, not xunit or xunit.core. If you do this wrong, you might have problems generating your NuGet package via dotnet pack.

Incidentally, this resolves an issue with upcoming changes to KoreBuild which uses the `IsTestProject` property to skip code signing for test projects.