ModelBuilding: Preserve Principal key configuration for FK on weak types
Issue: When an owned type is converted to weak type, we detach/attach relationships. In the process we are losing the principal key configuration for non-ownership Fks
This caused explicitly configured relationship to use shadow AK when PK type was different than FK.
Fix: We were resetting PrincipalKey when dependentType is not same as FK.DependentType. Though for owned to weak conversion it is never same but the condition is only supposed to capture inversion. Hence expanded the check to match if dependentType is same as PrincipalEntityType
Resolves #12423