Fix to #12738 - FirstOrDefault within Automapper causing Error
Problem was that the query had the same subquery present multiple times in the actual model and when we applied SubqueryMemberPushDownExpressionVisitor on one of those we would consequently mutate it in other place also, leading to invalid QM.
Previously we encountered this issue but only fixed it for pushdown via MemberExpression, missing cases that use EF.Property.

Fix is to clone the subquery before mutating it in the subquery member pushdown for EF.Property cases.