Fix to #12984 - Collection navigations should use reference equality by default
Problem was that when we were creating collection navigations we were using default ctors. This would cause (observable) HashSets to use element's Equals method when determining if the given entity is already part of the collection. This could lead to the same collection being added multiple times etc.

Fix is to special-case create collection delegates for well known concrete types like HashSet<T> and ObservableHashSet<T> and instantiate them using constructor that takes equality comparer argument. This way we can always ensure reference equality being perform on its elements.

For List<T> there is no such overload and for custom collection types we can't know which ctor to use so the behavior there stays the same.