Fix erroneous diff in data seeding with binary values
Fixes #13172

This change means we use structural comparisons for Equals now in regular DetectChanges. In the common case of DetectChanges, this short-circuits to quick same reference check, since the bytes are still not snapshotted. However, in case where the reference has changed, but the bytes are the same, we no longer get an erroneous Modified state, which as well as fixing this issue also means that we can avoid sending an update for data that hasn't changed.