Ensure QueryTrackingBehavior is reset when pooling
Fixes #13491

The issue was that if an option was not been changed from the default when the context instance was created, then it isn't saved away for resetting, which means that even if changed later it would not then be reset. The fix is to make sure these things are reset even if they were not changed when the instance was created.