Fix scaffolding type mapper FixedLength bug
Fixes #14160 Superceeds #14162

ScafoldingTypeMapper should have been passing `false` when looking for the default mapping, since, non-fixed length is the default. (Compare this to Unicode, where we pass `true` because Unicode is the default.)

Also included is a fix for test, which wasn't Asserting for fixed length, as found by Smit.

Also fixed the code in RelationalTypeMappingInfo that was for some reason using parsed-length to drive the fixed-length flag, although this was a secondary issue.