Start throwing by default for excessive service provider use
Issue #10236

Most of the changes here are two our tests where we use the internal service provider both as a convenience and to test specific behavior around the caching. The tests have been either:
* Updated to use an external service provider explicitly
* Updated to make use of a new option `ServiceProviderCachingEnabled`, which is on by default, but can be switched off to still use the internal service provider, but not to cache it.
* Left untouched where the behavior being tested depends on service provider caching.

This means our tests will not hit the error. We have a lot of testing of the general warning-as-error mechanism. Tested the new default throw here manually.