Add back registration for logging and memory cache
We were making sure these were registered in `AddDbContext` because EF used to need them to be registered, but now it's only optional depending on what you're doing. However, it seems that even in ASP.NET code this causes problems because people are using AddDbContext as a means to add these things to the service provider. This indicates poorly written tests/samples, but I think we can assume such will exist in the wild to, so adding these back in to avoid the break.