Add test for using ILoggerFactory with pool
This works because ILoggerFactory is a singleton service. Scoped services cannot be obtained correctly when options are registered as singleton, or pooling is used (which implies the same thing.) This is a limitation with pooling/singleton options which we always had and can keep--the fix is to not use pooling/singleton options.