Improve experience for HasOne/HasMany used with single string
Issue #9171

Now if these methods are called with a single string, then that string represents the navigation property name. If there is no CLR property with the given name, then an exception is thrown.

The exception is when the entity type has no CLR type, in which case the old behavior is preserved so as not to break existing model snapshots. Entity types with no CLR type are only valid in model snapshots.