Associate query loggers with the query context
Fixes #14923

This removes loggers from SelectExpression and RelationalCommand, since this can result in an old logger being used when query caching is in play. Instead, for query, the loggers are associated with the query context, which is new for each query.

This also alows some query/command services to move back to Singleton.