fix condition in CSharpDbContextGenerator.GenerateSequence
Correct the condition to append schema parameter to sequence creation code when generating DB context's OnModelCreating method. Even when sequence's schema is provide, schema parameter was not output to generated code. This change in the condition makes the code generation put sequence's schema to sequence creation code when it is provided.