Fix to #15720 - Query: Subquery Member pushdown
Problem was we were not handling a scenario with nested FirstOrDefault, e.g. customers.Select(c => orders.Select(orderDetails.FirstOrDefault()).FirstOrDefault().DetailName)
In this case we need to recursively pushdown the property until we don't encounter cardinality reducing operator.

Also resolves #15363 - Consecutive calls to FirstOrDefault() in a projection not translatable
This was already working but adding a test.