Use best-practices for Exception classes
Fixes #12285
Related #15214

It turns out that the code causing #15214 to happen is not due to normal serialization, and hence it is likely not possible to fix this in EF code. (See comments on the issue.) However, the issue (or similar) would still exist if the exception was serialized, and the fix is to use best practices for exceptions, so doing that.

For reference, see https://docs.microsoft.com/en-us/dotnet/standard/exceptions/best-practices-for-exceptions