Fix to #15778 - Query: null semantics visitor incorrectly classifies subquery returning single scalar value as non-nullable
Problem was that when we compared scalar subquery to null, we would assume the subquery is never null so we would never return any results. Scalar subqueries can be null when using entities.Where(e => false).Select(e => e.Name).FirstOrDefault() pattern.

Fix is to always assume SubSelectExpression is nullable.