Don't pull entity types from the context when building the Migrations history table
I think this is one of the root causes of the ASP.NET failures in https://github.com/aspnet/AspNetCore/pull/10939#issuecomment-500473365

The issue is that now that we discover DbSets from a context as a convention, re-using the convention-laden ModelBuilder for creating the history table also brings in types from the context.

For now, the fix is to remove this convention, but we may want to consider other options for building the history table.