Be explicit about type mappings being available on the finalized model
Part of #15663
Fixes #14433

The `GetTypeMapping` and `GetRelationalTypeMapping` methods will throw if there isn't a type mapping annotated on the model. Model finalization ensures that this is true.

Find methods still exist for cases where they need to be called before the model is finalized, and also so as not to break when not needed. However, see #16177