Implement context.Database.Log for simple SQL logging
Fixes #1199

Very similar to EF6, except the log level can also be set. For example:

```C#
context.Database.Log(Console.WriteLine);
```
or
```C#
context.Database.Log(Console.WriteLine, LogLevel.Information);
```
Only logs command execution for now, since that's all we have an interceptor for so far.

Uses the `EventData.ToString()` trick to get the same message formatting as we use for `ILogger`, but all the information is available to customize the output if needed.