Fix to #16076 - Query: Nav rewrite type mismatch
Problem was that when rewriting collection navigations into IQueryable, if the collection is composed on, we would always removing MaterializeCollectionNavigation. However, if the method was an instance method, e.g. List<T>.ToArray, we should preserve the original collection type, so that the method call is still possible to make.
Fix is to look at the declaring type of instance method, and only strip MaterializeCollectionNavigation if the instance type is IQueryable. For static methods, only remove MaterializeCollectionNavigation if the method is defined on Queryable or Enumerable