Fix cascade detach to actually cascade
Fixes #15960

The issue here is that when the first graph of Added entities are replaces by the second graph, it causes the first graph to be detached, but this was only happening at the first level of the graph.

Other test changes are because we were relying on being able to detach a principal without it cascading, even though cascades are configured.