Query: Add support for custom EntityQueryables
This makes look up for EntityQueryable in query tree bit more complex but it adds support for providers to add custom EntityQueryable to carry more metadata as needs to and allow rest of the pipeline to be transparent about it.
Current FromSql has be to converted during compilation phase since, we want to parameterize the arguments array in the methodCall.
If there is nothing to parameterize then providers can directly create custom Queryable.

Resolves #16326