Enable sql verification for #16086 - Query: subquery with order by but without Skip/Take gets the orderby removed
Sql we generate is not perfect (ordering on the joined set is not preserved) but there is not much we can do in the general case. We could improve this for some cases (outer set it's unambiguously sorted or is entity which we can sort ourselves).
Probably not worth given the marginal benefit.